`peerDependencies` --> `optionalDependencies`
