+++ .npmrc
@@ -1,6 +1,3 @@
-#Allow us to specify the version range of support for openmct in peer dependencies
-legacy-peer-deps=true
-
 loglevel=warn
 
 #Prevent folks from ignoring an important error when building from source
