+++ package.json
@@ -9,9 +9,9 @@
     "lint": "eslint src example",
     "lint:fix": "eslint src example --fix",
     "build:dist": "webpack --config webpack.prod.js",
-    "build:example": "npm install openmct@unstable --no-save --legacy-peer-deps=false",
+    "build:example": "npm install openmct@unstable --no-save",
     "build:example:master": "npm install nasa/openmct#master --no-save",
-    "postbuild:example": "node check-peer-dependencies.js",
+    "postbuild:example": "node check-optional-dependencies.js",
     "start": "npx webpack serve --config ./webpack.dev.js",
     "start:coverage": "npx webpack serve --config ./webpack.coverage.js",
     "prepare": "npm run build:dist",
@@ -30,7 +30,7 @@
   "engines": {
     "node": ">=16.19.1"
   },
-  "peerDependencies": {
+  "optionalDependencies": {
     "openmct": ">= 2.2.5"
   },
   "devDependencies": {
